refactor(schemes): move contracts to index.ts
