+++ src/core/auth.ts
@@ -6,7 +6,7 @@ import Router from '../utils/contracts/Router'
 import Scheme from '../schemes/Scheme'
 import TokenableScheme from '../schemes/TokenableScheme'
 import RefreshableScheme from '../schemes/RefreshableScheme'
-import SchemeCheck from '../schemes/contracts/SchemeCheck'
+import { SchemeCheck } from '../schemes'
 import Storage from './storage'
 import RedirectListener from './contracts/RedirectListener'
 import ErrorListener from './contracts/ErrorListener'
