--- src/schemes/contracts/SchemeCheck.ts
@@ -1,8 +1,0 @@
-export interface SchemeCheck {
-  valid: boolean
-  tokenExpired?: boolean
-  refreshTokenExpired?: boolean
-  isRefreshable?: boolean
-}
-
-export default SchemeCheck
