--- src/schemes/contracts/SchemePartialOptions.ts
@@ -1,9 +1,0 @@
-import { PartialExcept } from '../../types/utils'
-import SchemeOptions from './SchemeOptions'
-
-export type SchemePartialOptions<Options extends SchemeOptions> = PartialExcept<
-  Options,
-  keyof SchemeOptions
->
-
-export default SchemePartialOptions
